% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertListToDataFrame.R
\name{convertListToDataFrame}
\alias{convertListToDataFrame}
\title{convertListToDataFrame}
\usage{
convertListToDataFrame(listIN)
}
\arguments{
\item{listIN}{a list of IRF and IRFC, each list element contains two sub-elements named "mIRFdraw" and "mIRFCdraw"}
}
\value{
list of three elements
}
\description{
Converts list of IRF into a data.frame. Each list element corresponds to one IRF proposal.
}
\examples{

}
