% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateVAR.R
\name{estimateVAR}
\alias{estimateVAR}
\title{estimateVAR}
\usage{
estimateVAR(db, asy, asx, estmBeg, estmEnd, iIncpt = TRUE)
}
\arguments{
\item{db, }{time tibble: data}

\item{asy, }{string: names of LHS variates}

\item{asx, }{string: names of RHS variates}

\item{estmBeg, }{date: initial period in the estimation sample}

\item{estmEnd, }{date: last period in the estimation sample}

\item{iIncpt, }{logical: include intercept on the RHS, default = TRUE}
}
\value{
list
}
\description{
estimates a reduced-form VAR
}
