% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIRFunsigned.R
\name{getIRFunsigned}
\alias{getIRFunsigned}
\title{getIRFunsigned}
\usage{
getIRFunsigned(mComp, mSigma, h)
}
\arguments{
\item{mComp}{np x np matrix, companion form of the VAR(p) model}

\item{mSigma}{n x n covariance matrix of the reduced-form residuals}

\item{h}{a numeric sequence, 0,1,2,...,h length of IRF}
}
\value{
IRF data.frame (h + 1) x n^2
}
\description{
returns IRF proposal without any sign restrictions.
}
