% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderShocksInIRFprop.R
\name{orderShocksInIRFprop}
\alias{orderShocksInIRFprop}
\title{orderShocksInIRFprop}
\usage{
orderShocksInIRFprop(mIRF)
}
\arguments{
\item{mIRF}{(h+1) x 2 matrix of IRF proposal from the QR rotation of the B matrix with BB' = Sigma
(reduced-form residual covariance matrix).}
}
\value{
matrix of the same dimension in case shocks could be ordered or NULL in case
        shocks could not be identified
}
\description{
The function is meant to be used in a bivariate SVAR with two-demand and supply-shocks.
The function orders shocks based on the impact matrix in such a way that the supply shock
(+/-) followed by the demand shock (+/+).
}
\examples{

}
